Sun, Jul 7, 2024, 2:03 AM +05:30

+++ readability/readability.c
@@ -23,7 +23,7 @@ int main(void)
     /// Calculates Coleman-Liau index
     float L = (no_of_letters) / (no_of_words/100.0);
     float S = (no_of_sentences) / (no_of_words/100.0);
-    float index = 0.0588 * L - 0.296 * S - 15.8;
+    double index = 0.0588 * L - 0.296 * S - 15.8;
 
     // Printing grade level
     if (index < 0)
